restructor  repo to reduce top level files
