--- .babelrc
@@ -1,5 +1,0 @@
-{
-  "presets": ["@babel/preset-typescript", ["@babel/env"]],
-  "compact": true,
-  "plugins": ["lodash", "@babel/plugin-proposal-class-properties"]
-}
