+++ .npmignore
@@ -19,10 +19,8 @@ src
 *.yml
 Dockerfile
 *.png
-.babelrc
 .dockerignore
 .eslint*
 .prettierrc.js
-tsconfig.json
-webpack.config.babel.js
+tsconfig
 docs
