--- .prettierrc.js
@@ -1,13 +1,0 @@
-module.exports = {
-  singleQuote: true,
-  trailingComma: 'es5',
-  proseWrap: 'always',
-  overrides: [
-    {
-      files: ['*.js', '*.ts'],
-      options: {
-        printWidth: 80,
-      },
-    },
-  ],
-};
