+++ src/buffer.ts
