+++ src/client/options.ts
@@ -1,6 +1,6 @@
 import { isUndefined } from 'lodash';
 
-export default function loadOptions(): object {
+export function loadOptions(): object {
   const defaultOptions = { fontSize: 14 };
   try {
     return isUndefined(localStorage.options)
