--- src/client/verify.ts
@@ -1,4 +1,0 @@
-export default function verifyPrompt(e: { returnValue: string }): string {
-  e.returnValue = 'Are you sure?';
-  return e.returnValue;
-}
