+++ src/server/command/address.ts
@@ -1,8 +1,4 @@
-export default function address(
-  referer: string,
-  user: string,
-  host: string
-): string {
+export function address(referer: string, user: string, host: string): string {
   const match = referer.match('.+/ssh/([^/]+)$');
   const fallback = user ? `${user}@${host}` : host;
   return match ? `${match[1].split('?')[0]}@${host}` : fallback;
