+++ src/server/command/ssh/parse.ts
@@ -1,6 +1,6 @@
 import { isUndefined } from 'lodash';
 
-export default function parseCommand(command: string, path?: string): string {
+export function parseCommand(command: string, path?: string): string {
   if (command === 'login' && isUndefined(path)) return '';
   return !isUndefined(path)
     ? `$SHELL -c "cd ${path};${command === 'login' ? '$SHELL' : command}"`
