--- src/server/index.ts
@@ -1,4 +1,0 @@
-import WeTTy from './wetty';
-import init from './cli';
-
-export default { start: WeTTy, init };
