+++ src/server/shared/xterm.ts
