--- src/server/utils/index.ts
@@ -1,5 +1,0 @@
-import logger from './logger';
-
-export { logger };
-
-export * from './ssl';
