--- src/server/wetty/term/index.ts
@@ -1,5 +1,0 @@
-import spawn from './spawn';
-
-export { spawn };
-
-export * from './login';
