+++ src/shared/verify.ts
@@ -1,0 +1,4 @@
+export function verifyPrompt(e: { returnValue: string }): string {
+  e.returnValue = 'Are you sure?';
+  return e.returnValue;
+}
