--- tsconfig.json
@@ -1,13 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "es6",
-    "module": "commonjs",
-    "lib": ["es2015"],
-    "sourceMap": true,
-    "strict": true,
-    "moduleResolution": "node",
-    "typeRoots": ["node_module/@types"],
-    "outDir": "./dist"
-  },
-  "include": ["./src/**/*.ts"]
-}
