+++ tsconfig/esm.json
@@ -1,0 +1,14 @@
+{
+  "compilerOptions": {
+    "target": "ES2015",
+    "module": "ES2020",
+    "strict": true,
+    "sourceMap": true,
+    "esModuleInterop": true,
+    "declaration": true,
+    "forceConsistentCasingInFileNames": true,
+    "outDir": "../lib/esm",
+    "typeRoots": ["node_module/@types"]
+  },
+  "include": ["../src/**/*.ts"]
+}
